move testcases out from TestHoodieMergeOnReadTable.java
