+++ hudi-client/hudi-spark-client/src/test/java/org/apache/hudi/functional/SparkClientFunctionalTestSuite.java
@@ -25,7 +25,7 @@ import org.junit.platform.suite.api.SelectPackages;
 import org.junit.runner.RunWith;
 
 @RunWith(JUnitPlatform.class)
-@SelectPackages({"org.apache.hudi.client.functional", "org.apache.hudi.table"})
+@SelectPackages({"org.apache.hudi.client.functional", "org.apache.hudi.table.functional"})
 @IncludeTags("functional")
 public class SparkClientFunctionalTestSuite {
